Create rolling-saturn
For using in a server build lab or alpha testing environment, monitoring a terminal for new kernels being available and rpm build visual review.
If the output is too quick for you, adjust the sleep.
Run as super user. Current use is for keeping a little vm taking the updates as soon as it can from upstream and printing the results in a loop.